refactor(legacy): extract BudgetPlanConverter service from the command
Move the legacy-plan conversion logic out of the ConvertLegacyBudgetPlans
console command into a reusable App\Support\Budget\BudgetPlanConverter, so
it can also be invoked outside the console (the upcoming table->view swap
migration runs it). The command becomes a thin wrapper that handles the
dry-run transaction and console output via a logger callback. Behaviour is
unchanged (same preserved ids, same fingerprint); the conversion-helper
test now exercises the service directly.
